# api/openapi.yaml
openapi: 3.0.3
info:
  title: Nekopass API
  version: 0.14.1
  description: HTTP backend; optional HTTPS termination at a trusted reverse proxy. Admin-created accounts. JSON requests, same-origin browser access. Byte counters combine both TCP directions once.
servers:
  - url: /api/v1
security:
  - session: []
  - httpSession: []
paths:
  /admin/updates/panel:
    get:
      summary: Read panel update support and task status (admin only)
      responses:
        '200': {description: Current version, helper availability and latest task}
    post:
      summary: Queue upgrade to the latest official release (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [version]
              properties:
                version: {type: string, pattern: '^v[0-9]+\.[0-9]+\.[0-9]+$'}
      responses:
        '202': {description: Durable update task queued for native systemd updater}
        '409': {description: No newer stable release, helper unavailable or another update is pending}
  /admin/updates:
    get:
      summary: Administrator only; check latest stable GitHub release
      responses:
        '200': {description: 'current_version, latest (version, url, published_at), update_available. Successful checks cached for 5 minutes; errors for 1 minute.'}
        '403': {$ref: '#/components/responses/Error'}
        '502': {description: GitHub unavailable or release files incomplete}
  /admin/nodes/{id}/update:
    parameters:
      - {name: id, in: path, required: true, schema: {type: integer, format: int64, minimum: 1}}
    post:
      summary: Administrator only; enqueue latest stable release for this online Agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [version]
              properties:
                version: {type: string, example: v0.10.0}
      responses:
        '200': {description: 'generation, version and queued state; Agent protocol 9 and installed updater required. Update progress appears in admin nodes only. Uses fixed hajidishu/nekopass Releases; preserves node identity and quota state.'}
        '403': {$ref: '#/components/responses/Error'}
        '409': {description: Offline/legacy node, pending task or version conflict}
  /admin/nodes/{id}/ddns:
    parameters:
      - {name: id, in: path, required: true, schema: {type: integer, format: int64, minimum: 1}}
    get:
      summary: Administrator only; node DDNS settings and latest synchronization status
      responses:
        '200': {description: 'node_id, node_name, protocol_version, online, config (including API token), generation, status and status_received_at. Never exposed to user APIs or other nodes.'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
    put:
      summary: Administrator only; save Cloudflare DDNS and push to this node
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                enabled: {type: boolean, default: false}
                provider: {type: string, enum: [cloudflare]}
                record_name: {type: string, example: node.example.com}
                token: {type: string, description: Cloudflare API token with DNS Edit permission}
                zone_id: {type: string, description: '32 hex characters; blank discovers automatically and requires Zone Read permission'}
                ipv4: {type: boolean, default: true}
                ipv6: {type: boolean, default: false}
                interval_seconds: {type: integer, minimum: 60, maximum: 86400, default: 300}
                ttl: {type: integer, description: '1 for automatic or 60–86400 seconds', default: 300}
                ipv4_url: {type: string, default: 'https://api.ipify.org'}
                ipv6_url: {type: string, default: 'https://api6.ipify.org'}
                use_for_ingress: {type: boolean, default: true}
                use_for_egress: {type: boolean, default: true}
      responses:
        '200': {description: 'generation; state pending until Agent protocol 8 or newer reports. Enabled address binding replaces selected node addresses with record_name. Disabling preserves records and node addresses.'}
        '400': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
  /admin/nodes/{id}/ddns/run:
    parameters:
      - {name: id, in: path, required: true, schema: {type: integer, format: int64, minimum: 1}}
    post:
      summary: Administrator only; request immediate reconciliation of enabled DDNS
      responses:
        '200': {description: generation}
        '403': {$ref: '#/components/responses/Error'}
        '409': {description: Save and enable DDNS first}
  /wallet:
    get:
      summary: Current account balance and most recent 100 wallet entries
      responses:
        '200': {description: 'balance_cents integer, currency CNY, entries, minimum_recharge_cents, empty payment_channels until a gateway is implemented'}
        '401': {$ref: '#/components/responses/Error'}
  /orders:
    get:
      summary: Current account orders only, including manual recharge and balance purchases
      parameters:
        - {name: page, in: query, schema: {type: integer, minimum: 1, maximum: 100000, default: 1}}
      responses:
        '200': {description: 'items, total and page; 20 rows per page; historical plan names and purchase snapshots retained'}
...
  /shop/quote:
    post:
      summary: Preview balance purchase or renewal dates and price without charging
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [plan_id, cycle]
              properties:
                plan_id: {type: integer, format: int64}
                cycle: {$ref: '#/components/schemas/BillingCycle'}
      responses:
        '200': {description: 'plan_id, plan_name, cycle, amount_cents, balance_cents, expires_at, next_reset_at, epoch, kind and replaces_plan'}
        '400': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
  /shop/purchase:
    post:
      summary: Atomically debit wallet, record order and start a new quota epoch; same request_key replays without charging
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [plan_id, cycle, request_key, expected_price_cents, expected_epoch]
              properties:
                plan_id: {type: integer, format: int64}
                cycle: {$ref: '#/components/schemas/BillingCycle'}
                request_key: {type: string, pattern: '^[A-Za-z0-9_-]{16,80}$'}
                expected_price_cents: {type: integer, format: int64, minimum: 0, description: Price returned by quote}
                expected_epoch: {type: integer, format: int64, minimum: 0, description: Epoch returned by quote; stale concurrent confirmations are rejected}
      responses:
        '200': {description: 'id, balance_cents, expires_at and next_reset_at; replayed true for an existing request'}
        '409': {description: Insufficient funds, stale quote, conflicting request key or rule limit exceeded}
  /admin/users/{id}/recharge:
    parameters:
      - {$ref: '#/components/parameters/ID'}
    post:
      security: [{session: []}, {httpSession: []}, {adminKey: []}]
      summary: Audited administrator recharge; does not invoke any payment gateway
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount, note, request_key]
              properties:
                amount: {type: string, example: '100.00', description: Positive CNY amount with at most two decimal places}
                note: {type: string, minLength: 1, maxLength: 300}
                request_key: {type: string, pattern: '^[A-Za-z0-9_-]{16,80}$'}
      responses:
        '200': {description: 'id and balance_cents; identical replay does not credit twice'}
        '400': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
...
  /admin/settings:
    get:
      security: [{session: []}, {httpSession: []}, {adminKey: []}]
      summary: Read system settings and whether a management API key is configured
      responses:
        '200':
          description: No plaintext key or key hash is returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  settings: {$ref: '#/components/schemas/SystemSettings'}
                  api_key_configured: {type: boolean}
        '403': {$ref: '#/components/responses/Error'}
    put:
      security: [{session: []}, {httpSession: []}, {adminKey: []}]
      summary: Replace system settings; public endpoints do not rebind server listeners
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SystemSettings'}
      responses:
        '200': {description: Saved settings}
        '400': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
  /admin/settings/api-key:
    post:
      security: [{session: []}, {httpSession: []}, {adminKey: []}]
      summary: Generate or replace the management API key; plaintext returned only once
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                key: {type: string, description: Empty generates a key; custom keys require 32–256 non-whitespace characters}
      responses:
        '200': {description: Object with key string; store securely}
        '400': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
    delete:
      security: [{session: []}, {httpSession: []}, {adminKey: []}]
      summary: Revoke the management API key without affecting node keys
      responses:
        '200': {description: Revoked}
        '403': {$ref: '#/components/responses/Error'}
  /admin/nodes/{id}/install-command:
    parameters:
      - {$ref: '#/components/parameters/ID'}
    post:
      security: [{session: []}, {httpSession: []}, {adminKey: []}]
      summary: Generate a reusable wget and bash command with explicit node connection parameters
      responses:
        '200':
          description: Includes the persistent node key, panel URL, HTTP/HTTPS gRPC endpoint and release source/version. HTTPS verifies system-trusted certificates; HTTP uses test h2c. No custom control CA. Administrator only; does not rotate the node key.
          content:
            application/json:
              schema:
                type: object
                properties:
                  command: {type: string}
                  expires_at: {type: string, nullable: true, description: Null for persistent-key commands}
                  existing_node: {type: boolean, description: Bound nodes require upgrade on the original host}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
  /node-install/redeem:
    post:
      security: [{installTicket: []}]
      summary: Atomically redeem a one-use installation credential
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                current_token: {type: string, description: Required for already-bound nodes; exact existing node key}
      responses:
        '200':
          description: Private connection and release configuration
          content:
            application/json:
              schema:
                type: object
                properties:
                  node_id: {type: integer, format: int64}
                  server: {type: string}
                  token: {type: string}
                  version: {type: string}
                  download_base: {type: string}
        '401': {description: Invalid, expired or already-used ticket}
        '409': {description: Bound node identity mismatch}
...
  /admin/announcement:
    get:
      summary: Admin reads the editable site announcement
      responses:
        '200': {description: Object with content string}
        '403': {$ref: '#/components/responses/Error'}
    put:
      summary: Admin publishes the site announcement
      requestBody: {$ref: '#/paths/~1announcement/put/requestBody'}
      responses:
        '200': {description: Saved announcement}
  /admin/users:
    get:
      summary: Admin lists user accounts, plans and usage for the management area
      responses:
        '200': {description: User array}
        '403': {$ref: '#/components/responses/Error'}
    post:
      summary: Admin creates a non-admin account
      requestBody: {$ref: '#/components/requestBodies/User'}
      responses:
        '200': {description: Created user ID}
  /admin/users/{userID}:
    parameters:
      - name: userID
        in: path
        required: true
        schema: {type: integer, format: int64}
    get:
      summary: Admin loads a specific account for its dedicated rules page
      responses:
        '200': {description: Single-element user resource array}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
    put:
      summary: Admin edits account and plan assignment
      requestBody: {$ref: '#/components/requestBodies/User'}
      responses:
        '200': {description: Updated user ID}
  /login:
    post:
      security: []
      summary: Login and set an HttpOnly session cookie; Secure follows the trusted request protocol
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: {type: string}
                password: {type: string}
      responses:
        '200': {description: 'id, username, is_admin; sets cookie'}
        '401': {$ref: '#/components/responses/Error'}
        '429': {$ref: '#/components/responses/Error'}
...
  /users:
    get:
      summary: All accounts for admin, own account otherwise
      responses:
        '200': {description: 'Array including traffic_bytes, charged_bytes, allocated_bytes and user settings; password hashes are never returned'}
    post:
      summary: Admin creates a non-admin user
      requestBody: {$ref: '#/components/requestBodies/User'}
      responses:
        '200': {description: Created user ID}
        '403': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
  /users/{id}:
    parameters:
      - {$ref: '#/components/parameters/ID'}
    put:
      summary: Admin updates account and plan assignment; empty password preserves it; direct limit overrides are rejected
      requestBody: {$ref: '#/components/requestBodies/User'}
      responses:
        '200': {description: Updated user ID}
        '409': {$ref: '#/components/responses/Error'}
  /nodes:
    get:
      summary: Admin-only legacy alias for node management; users must use node-status or rule-nodes
      responses:
        '200': {description: Node array}
    post:
      summary: Admin creates node; its stable token remains visible in node management
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: {type: string, maxLength: 80}
      responses:
        '200': {description: 'id, token'}
  /rules:
    get:
      summary: Current account rules only, including administrators
      responses:
        '200': {description: Rule array including node_name, username, traffic_bytes}
    post:
      summary: Create TCP rule on an authorized ingress node, optionally via an authorized plaintext exit
      requestBody: {$ref: '#/components/requestBodies/Rule'}
      responses:
        '200': {description: Created rule ID}
        '409': {$ref: '#/components/responses/Error'}
  /rules/{id}:
    parameters:
      - {$ref: '#/components/parameters/ID'}
    put:
      summary: Update owned rule; owner cannot change; switching node requires authorization
      requestBody: {$ref: '#/components/requestBodies/Rule'}
      responses:
        '200': {description: Updated rule ID}
    delete:
      summary: Delete owned rule and stop existing connections on next synchronization
      responses:
        '200': {description: Deleted}
        '404': {$ref: '#/components/responses/Error'}
  /announcement:
    get:
      summary: Read site announcement as plain text
      responses:
        '200': {description: 'Object with content string'}
    put:
      summary: Admin updates announcement (20000 UTF-8 bytes maximum)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content: {type: string}
      responses:
        '200': {description: Saved announcement}
  /rule-groups:
    get:
      summary: List accessible groups with user_id and rule_count
      responses:
        '200': {description: Array of groups}
    post:
      summary: Create group; ordinary users always own their groups
      requestBody: {$ref: '#/components/requestBodies/Group'}
      responses:
        '200': {description: Group ID}
  /rule-groups/{id}:
    parameters:
      - {$ref: '#/components/parameters/ID'}
    put:
      summary: Rename accessible group
      requestBody: {$ref: '#/components/requestBodies/Group'}
      responses:
        '200': {description: Group ID}
    delete:
      summary: Delete group and move its rules to ungrouped
      responses:
        '200': {description: Deleted}
  /nodes/{id}:
    parameters:
      - {$ref: '#/components/parameters/ID'}
    put:
      summary: Admin changes node name and displayed public address
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, public_address]
              properties:
                name: {type: string}
                public_address: {type: string, description: IP or hostname without scheme or port; empty allowed}
      responses:
        '200': {description: Updated}
  /rules/batch:
    post:
      summary: Atomically mutate 1–500 accessible rules; all ownership checks precede writes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ids, action]
              properties:
                ids: {type: array, minItems: 1, maxItems: 500, uniqueItems: true, items: {type: integer, format: int64}}
                action: {type: string, enum: [enable, disable, delete, clear, group, switch]}
                group_id: {type: integer, format: int64, description: For group action; 0 means ungrouped}
                node_id: {type: integer, format: int64, description: For switch action; authorized for all selected users; random ports assigned}
      responses:
        '200': {description: 'ok and affected count; clear only resets display baseline, never quota or history'}
        '403': {$ref: '#/components/responses/Error'}
  /rules/import:
    post:
      summary: Atomic import; any invalid rule rolls back the complete batch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rules]
              properties:
                rules:
                  type: array
                  minItems: 1
                  maxItems: 500
                  items: {$ref: '#/components/schemas/RuleInput'}
      responses:
        '200': {description: Object with created ids array}
  /rules/stats:
    get:
      summary: Actual traffic over last 24 hours grouped by hour; access restricted to current user's rules
      parameters:
        - name: rule_id
          in: query
          schema: {type: integer, format: int64}
        - name: user_id
          in: query
          description: Deprecated; ignored. Scope comes from session or admin-managed path
          schema: {type: integer, format: int64}
      responses:
        '200': {description: Array of time (timestamp) and bytes (int64); scope is current account or the explicit admin-managed user}
...
components:
  securitySchemes:
    session: {type: apiKey, in: cookie, name: nekopass_session, description: HTTPS session}
    httpSession: {type: apiKey, in: cookie, name: nekopass_session_http, description: HTTP session; isolated from HTTPS cookies}
    adminKey:
      type: http
      scheme: bearer
      description: Optional alternative to session for all /admin/* API routes only. Owner must remain an enabled administrator. Does not authenticate HTML pages.
    installTicket:
      type: http
      scheme: bearer
      description: Short-lived one-use node-specific credential; not a management API key
  parameters:
    ID: {name: id, in: path, required: true, schema: {type: integer, format: int64}}
  responses:
    Error:
      description: JSON error
      content:
        application/json:
          schema:
            type: object
            properties:
              error: {type: string}
  requestBodies:
...
    User:
      required: true
      content:
        application/json:
          schema: {$ref: '#/components/schemas/UserInput'}
    Rule:
      required: true
      content:
        application/json:
          schema: {$ref: '#/components/schemas/RuleInput'}
  schemas:
    BillingCycle:
      type: string
      enum: [monthly, quarterly, semiannual, annual, biennial, triennial, onetime]
...